+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.13a
+- new submodule "input" to get input from users
+- new function `input.reply` to get specific reply from user
+
 # 0.0.12a1
 - checks.dm_from_this_guild now accepts aditional argument :delete: to remove the message if called from a guild. This will work only if it comes from the specified guild. If its a different guild the check will just return False